chore: move `makeCounter` to utils for reuse
