+++ lib/cmap/connection_pool.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const EventEmitter = require('events').EventEmitter;
-const makeCounter = require('./counter').makeCounter;
+const makeCounter = require('../util').makeCounter;
 const Connection = require('./connection').CMAPConnection;
 const WaitQueue = require('./wait_queue').WaitQueue;
 const ConnectionManager = require('./connection_manager').ConnectionManager;
