+++ lib/utils.js
@@ -685,6 +685,15 @@ class MongoDBNamespace {
   }
 }
 
+function* makeCounter(seed) {
+  let count = seed || 0;
+  while (true) {
+    const newCount = count;
+    count += 1;
+    yield newCount;
+  }
+}
+
 module.exports = {
   filterOptions,
   mergeOptions,
@@ -713,5 +722,6 @@ module.exports = {
   SUPPORTS,
   MongoDBNamespace,
   resolveReadPreference,
-  emitDeprecationWarning
+  emitDeprecationWarning,
+  makeCounter
 };
